+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### jannovar-statistics
 
+* Also counting number of variants on contigs
 * Fixing counting bug that made UTR3 variants be counted as UTR5
 * Fixing NPE in case of null variant annotations (e.g., unknown contig)